[proliferate] split playground composer
