+++ desktop/src/components/playground/PlaygroundComposerActions.tsx
@@ -1,0 +1,6 @@
+export const noop = () => {};
+
+export const noopAsync = async () => {};
+
+export const revealExampleUrl = async () =>
+  "https://accounts.example.com/oauth/authorize?client_id=redacted";
