+++ scripts/max_lines_allowlist.txt
@@ -33,7 +33,6 @@ anyharness/sdk/src/reducer/transcript.ts
 anyharness/sdk/src/reducer/__tests__/transcript.test.ts
 anyharness/sdk-react/src/hooks/sessions.ts
 desktop/src/components/ui/icons.tsx
-desktop/src/components/playground/PlaygroundComposer.tsx
 desktop/src/components/plugins/detail/ConnectorDetailModal.tsx
 desktop/src/components/settings/panes/OrganizationPane.tsx
 desktop/src/components/settings/panes/review/ReviewDefaultsSection.tsx
